fix : fix integer type overflow
change totalPurchaseAmount & totalReward type from int to long in LottoResult

+++ src/main/java/lotto/result/LottoResult.java
@@ -14,9 +14,9 @@ public class LottoResult {
 
     private int totalLottoCount;
 
-    private int totalPurchaseAmount;
+    private long totalPurchaseAmount;
 
-    private int totalReward;
+    private long totalReward;
 
     private String message;
